ad_server/requirements.txt
captcha==0.3
certifi==2021.10.8
cffi==1.15.0
click==8.0.3
cloudinary==1.28.0
cryptography==36.0.0
Flask==2.0.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
Pillow==8.4.0
psycopg2-binary==2.9.2
pycparser==2.21
python-dotenv==0.19.2
six==1.16.0
SQLAlchemy==1.4.27
typing_extensions==4.0.1
urllib3==1.26.7
Werkzeug==2.0.2
zipp==3.6.0
